backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
requests>=2.31.0
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.0.0
drf-spectacular>=0.27.0
bleach>=6.0.0
